actually upload vbox guest additions to the box, but dont delete them

+++ packer/scripts/centos/cleanup.sh
@@ -1,6 +1,8 @@
 #!/bin/bash -eux
 yum -y clean all
-rm -rf VBoxGuestAdditions_*.iso VBoxGuestAdditions_*.iso.?
+if [ $VM_TOOLS != 'disable' ]; then
+  rm -rf VBoxGuestAdditions_*.iso VBoxGuestAdditions_*.iso.?
+fi
 rm -f /tmp/chef*rpm
 
 # clean up redhat interface persistence
